Feat: Refactor parameters in messageLayer (#1039)
* Feat: add parameter utils

* Refactor: changed more params

* Refactor: started adding statement cli parameters

* Refactor: removed test

* Refactor: added more statement parameters

* Refactor: removed more parameters

* Feat: updated to latest hive.go

* Feat: go mod tidy

* Feat: started adding comments back

* Feat: added doc comments to parameters

* Feat: updated hive.go

* Feat: updated hive.go

* Feat: removed logger

* Refactor: changed more loggers

* Feat: updated further parameters

* Refactor: removed last log

* Refactor: refactored files to be 1 file per plugin
