+++ plugins/messagelayer/consensus_plugin.go
@@ -1,128 +1,79 @@
 package messagelayer
 
 import (
+	"context"
+	"fmt"
 	"net"
 	"strconv"
 	"sync"
 	"time"
 
+	"github.com/iotaledger/goshimmer/packages/consensus/fcob"
+	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/packages/metrics"
 	"github.com/iotaledger/goshimmer/packages/prng"
 	"github.com/iotaledger/goshimmer/packages/shutdown"
+	"github.com/iotaledger/goshimmer/packages/tangle"
 	"github.com/iotaledger/goshimmer/packages/vote"
 	"github.com/iotaledger/goshimmer/packages/vote/fpc"
 	votenet "github.com/iotaledger/goshimmer/packages/vote/net"
 	"github.com/iotaledger/goshimmer/packages/vote/opinion"
 	"github.com/iotaledger/goshimmer/packages/vote/statement"
+	"github.com/iotaledger/goshimmer/plugins/autopeering"
 	"github.com/iotaledger/goshimmer/plugins/autopeering/local"
-	"github.com/iotaledger/goshimmer/plugins/config"
+	"github.com/iotaledger/goshimmer/plugins/clock"
+	"github.com/iotaledger/goshimmer/plugins/remotelog"
+	"github.com/iotaledger/hive.go/autopeering/peer"
 	"github.com/iotaledger/hive.go/autopeering/peer/service"
 	"github.com/iotaledger/hive.go/daemon"
 	"github.com/iotaledger/hive.go/events"
-	"github.com/iotaledger/hive.go/logger"
+	"github.com/iotaledger/hive.go/identity"
 	"github.com/iotaledger/hive.go/node"
-	flag "github.com/spf13/pflag"
+	"google.golang.org/grpc"
+	"google.golang.org/protobuf/proto"
 )
 
-const (
-	// ConsensusPluginName contains the human readable name of the plugin.
-	ConsensusPluginName = "Consensus"
-
-	// CfgFPCQuerySampleSize defines how many nodes will be queried each round.
-	CfgFPCQuerySampleSize = "fpc.querySampleSize"
-
-	// CfgFPCRoundInterval defines how long a round lasts (in seconds)
-	CfgFPCRoundInterval = "fpc.roundInterval"
-
-	// CfgFPCListen defines if the FPC service should listen.
-	CfgFPCListen = "fpc.listen"
-
-	// CfgFPCBindAddress defines on which address the FPC service should listen.
-	CfgFPCBindAddress = "fpc.bindAddress"
-
-	// CfgWaitForStatement is the time in seconds for which the node wait for receiving the new statement.
-	CfgWaitForStatement = "statement.waitForStatement"
-
-	// CfgManaThreshold defines the Mana threshold to accept/write a statement.
-	CfgManaThreshold = "statement.manaThreshold"
-
-	// CfgCleanInterval defines the time interval [in minutes] for cleaning the statement registry.
-	CfgCleanInterval = "statement.cleanInterval"
-
-	// CfgDeleteAfter defines the time [in minutes] after which older statements are deleted from the registry.
-	CfgDeleteAfter = "statement.deleteAfter"
-	// CfgWriteStatement defines if the node should write statements.
-	CfgWriteStatement = "statement.writeStatement"
-)
-
-func init() {
-	flag.Bool(CfgFPCListen, true, "if the FPC service should listen")
-	flag.Bool(CfgWriteStatement, false, "if the node should make statements")
-	flag.Int(CfgFPCQuerySampleSize, 21, "Size of the voting quorum (k)")
-	flag.Int64(CfgFPCRoundInterval, 10, "FPC round interval [s]")
-	flag.String(CfgFPCBindAddress, "0.0.0.0:10895", "the bind address on which the FPC vote server binds to")
-	flag.Int(CfgWaitForStatement, 5, "the time in seconds for which the node wait for receiving the new statement")
-	flag.Float64(CfgManaThreshold, 1., "Mana threshold to accept/write a statement")
-	flag.Int(CfgCleanInterval, 5, "the time in minutes after which the node cleans the statement registry")
-	flag.Int(CfgDeleteAfter, 5, "the time in minutes after which older statements are deleted from the registry")
-}
+// region Plugin ///////////////////////////////////////////////////////////////////////////////////////////////////////
 
 var (
 	// plugin is the plugin instance of the statement plugin.
-	consensusPlugin      *node.Plugin
-	consensusPluginOnce  sync.Once
-	voter                *fpc.FPC
-	voterOnce            sync.Once
-	voterServer          *votenet.VoterServer
-	roundIntervalSeconds int64
-	consensusPluginLog   *logger.Logger
-	registry             *statement.Registry
-	registryOnce         sync.Once
-	waitForStatement     int
-	listen               bool
-	cleanInterval        int
-	deleteAfter          int
-	writeStatement       bool
+	consensusPlugin     *node.Plugin
+	consensusPluginOnce sync.Once
+	voter               *fpc.FPC
+	voterOnce           sync.Once
+	voterServer         *votenet.VoterServer
+	registry            *statement.Registry
+	registryOnce        sync.Once
 )
 
 // ConsensusPlugin returns the consensus plugin.
 func ConsensusPlugin() *node.Plugin {
 	consensusPluginOnce.Do(func() {
-		consensusPlugin = node.NewPlugin(ConsensusPluginName, node.Enabled, configureConsensusPlugin, runConsensusPlugin)
+		consensusPlugin = node.NewPlugin("Consensus", node.Enabled, configureConsensusPlugin, runConsensusPlugin)
 	})
 	return consensusPlugin
 }
 
-func configureConsensusPlugin(*node.Plugin) {
-	consensusPluginLog = logger.NewLogger(ConsensusPluginName)
-
+func configureConsensusPlugin(plugin *node.Plugin) {
 	configureRemoteLogger()
-
-	roundIntervalSeconds = config.Node().Int64(CfgFPCRoundInterval)
-	waitForStatement = config.Node().Int(CfgWaitForStatement)
-	listen = config.Node().Bool(CfgFPCListen)
-	cleanInterval = config.Node().Int(CfgCleanInterval)
-	deleteAfter = config.Node().Int(CfgDeleteAfter)
-	writeStatement = config.Node().Bool(CfgWriteStatement)
-
-	configureFPC()
+	configureFPC(plugin)
 
 	// subscribe to FCOB events
 	ConsensusMechanism().Events.Vote.Attach(events.NewClosure(func(id string, initOpn opinion.Opinion) {
 		if err := Voter().Vote(id, vote.ConflictType, initOpn); err != nil {
-			consensusPluginLog.Warnf("FPC vote: %s", err)
+			plugin.LogWarnf("FPC vote: %s", err)
 		}
 	}))
 	ConsensusMechanism().Events.Error.Attach(events.NewClosure(func(err error) {
-		consensusPluginLog.Errorf("FCOB error: %s", err)
+		plugin.LogErrorf("FCOB error: %s", err)
 	}))
 
 	// subscribe to message-layer
 	Tangle().ConsensusManager.Events.MessageOpinionFormed.Attach(events.NewClosure(readStatement))
 }
 
-func runConsensusPlugin(*node.Plugin) {
-	runFPC()
+func runConsensusPlugin(plugin *node.Plugin) {
+	runFPC(plugin)
 }
 
 // Voter returns the DRNGRoundBasedVoter instance used by the FPC plugin.
@@ -141,55 +92,54 @@ func Registry() *statement.Registry {
 	return registry
 }
 
-func configureFPC() {
-	if listen {
+func configureFPC(plugin *node.Plugin) {
+	if FPCParameters.Listen {
 		lPeer := local.GetInstance()
-		bindAddr := config.Node().String(CfgFPCBindAddress)
-		_, portStr, err := net.SplitHostPort(bindAddr)
+		_, portStr, err := net.SplitHostPort(FPCParameters.BindAddress)
 		if err != nil {
-			consensusPluginLog.Fatalf("FPC bind address '%s' is invalid: %s", bindAddr, err)
+			plugin.LogFatalf("FPC bind address '%s' is invalid: %s", FPCParameters.BindAddress, err)
 		}
 		port, err := strconv.Atoi(portStr)
 		if err != nil {
-			consensusPluginLog.Fatalf("FPC bind address '%s' is invalid: %s", bindAddr, err)
+			plugin.LogFatalf("FPC bind address '%s' is invalid: %s", FPCParameters.BindAddress, err)
 		}
 
 		if err := lPeer.UpdateService(service.FPCKey, "tcp", port); err != nil {
-			consensusPluginLog.Fatalf("could not update services: %v", err)
+			plugin.LogFatalf("could not update services: %v", err)
 		}
 	}
 
 	Voter().Events().RoundExecuted.Attach(events.NewClosure(func(roundStats *vote.RoundStats) {
-		if writeStatement {
+		if StatementParameters.WriteStatement {
 			makeStatement(roundStats)
 		}
 		peersQueried := len(roundStats.QueriedOpinions)
 		voteContextsCount := len(roundStats.ActiveVoteContexts)
-		consensusPluginLog.Debugf("executed round with rand %0.4f for %d vote contexts on %d peers, took %v", roundStats.RandUsed, voteContextsCount, peersQueried, roundStats.Duration)
+		plugin.LogDebugf("executed round with rand %0.4f for %d vote contexts on %d peers, took %v", roundStats.RandUsed, voteContextsCount, peersQueried, roundStats.Duration)
 	}))
 
 	Voter().Events().Finalized.Attach(events.NewClosure(ConsensusMechanism().ProcessVote))
 	Voter().Events().Finalized.Attach(events.NewClosure(func(ev *vote.OpinionEvent) {
 		if ev.Ctx.Type == vote.ConflictType {
-			consensusPluginLog.Infof("FPC finalized for transaction with id '%s' - final opinion: '%s'", ev.ID, ev.Opinion)
+			plugin.LogInfof("FPC finalized for transaction with id '%s' - final opinion: '%s'", ev.ID, ev.Opinion)
 		}
 	}))
 
 	Voter().Events().Failed.Attach(events.NewClosure(func(ev *vote.OpinionEvent) {
 		if ev.Ctx.Type == vote.ConflictType {
-			consensusPluginLog.Warnf("FPC failed for transaction with id '%s' - last opinion: '%s'", ev.ID, ev.Opinion)
+			plugin.LogWarnf("FPC failed for transaction with id '%s' - last opinion: '%s'", ev.ID, ev.Opinion)
 		}
 	}))
 
 }
 
-func runFPC() {
+func runFPC(plugin *node.Plugin) {
 	const ServerWorkerName = "FPCVoterServer"
 
-	if listen {
+	if FPCParameters.Listen {
 		if err := daemon.BackgroundWorker(ServerWorkerName, func(shutdownSignal <-chan struct{}) {
 			stopped := make(chan struct{})
-			bindAddr := config.Node().String(CfgFPCBindAddress)
+			bindAddr := FPCParameters.BindAddress
 			voterServer = votenet.New(Voter(), OpinionRetriever, bindAddr,
 				metrics.Events().FPCInboundBytes,
 				metrics.Events().FPCOutboundBytes,
@@ -197,9 +147,9 @@ func runFPC() {
 			)
 
 			go func() {
-				consensusPluginLog.Infof("%s started, bind-address=%s", ServerWorkerName, bindAddr)
+				plugin.LogInfof("%s started, bind-address=%s", ServerWorkerName, bindAddr)
 				if err := voterServer.Run(); err != nil {
-					consensusPluginLog.Errorf("Error serving: %s", err)
+					plugin.LogErrorf("Error serving: %s", err)
 				}
 				close(stopped)
 			}()
@@ -210,18 +160,18 @@ func runFPC() {
 			case <-stopped:
 			}
 
-			consensusPluginLog.Infof("Stopping %s ...", ServerWorkerName)
+			plugin.LogInfof("Stopping %s ...", ServerWorkerName)
 			voterServer.Shutdown()
-			consensusPluginLog.Infof("Stopping %s ... done", ServerWorkerName)
+			plugin.LogInfof("Stopping %s ... done", ServerWorkerName)
 		}, shutdown.PriorityFPC); err != nil {
-			consensusPluginLog.Panicf("Failed to start as daemon: %s", err)
+			plugin.Panicf("Failed to start as daemon: %s", err)
 		}
 	}
 
 	if err := daemon.BackgroundWorker("FPCRoundsInitiator", func(shutdownSignal <-chan struct{}) {
-		consensusPluginLog.Infof("Started FPC round initiator")
-		defer consensusPluginLog.Infof("Stopped FPC round initiator")
-		unixTsPRNG := prng.NewUnixTimestampPRNG(roundIntervalSeconds)
+		plugin.LogInfof("Started FPC round initiator")
+		defer plugin.LogInfof("Stopped FPC round initiator")
+		unixTsPRNG := prng.NewUnixTimestampPRNG(FPCParameters.RoundInterval)
 		unixTsPRNG.Start()
 		defer unixTsPRNG.Stop()
 	exit:
@@ -229,31 +179,339 @@ func runFPC() {
 			select {
 			case r := <-unixTsPRNG.C():
 				if err := voter.Round(r); err != nil {
-					consensusPluginLog.Warnf("unable to execute FPC round: %s", err)
+					plugin.LogWarnf("unable to execute FPC round: %s", err)
 				}
 			case <-shutdownSignal:
 				break exit
 			}
 		}
 	}, shutdown.PriorityFPC); err != nil {
-		consensusPluginLog.Panicf("Failed to start as daemon: %s", err)
+		plugin.Panicf("Failed to start as daemon: %s", err)
 	}
 
 	if err := daemon.BackgroundWorker("StatementCleaner", func(shutdownSignal <-chan struct{}) {
-		consensusPluginLog.Infof("Started Statement Cleaner")
-		defer consensusPluginLog.Infof("Stopped Statement Cleaner")
-		ticker := time.NewTicker(time.Duration(cleanInterval) * time.Minute)
+		plugin.LogInfof("Started Statement Cleaner")
+		defer plugin.LogInfof("Stopped Statement Cleaner")
+		ticker := time.NewTicker(time.Duration(StatementParameters.CleanInterval) * time.Minute)
 		defer ticker.Stop()
 	exit:
 		for {
 			select {
 			case <-ticker.C:
-				Registry().Clean(time.Duration(deleteAfter) * time.Minute)
+				Registry().Clean(time.Duration(StatementParameters.DeleteAfter) * time.Minute)
 			case <-shutdownSignal:
 				break exit
 			}
 		}
 	}, shutdown.PriorityFPC); err != nil {
-		consensusPluginLog.Panicf("Failed to start as daemon: %s", err)
+		plugin.Panicf("Failed to start as daemon: %s", err)
 	}
 }
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region OpinionGivers ////////////////////////////////////////////////////////////////////////////////////////////////
+
+// OpinionGiver is a wrapper for both statements and peers.
+type OpinionGiver struct {
+	id   identity.ID
+	view *statement.View
+	pog  *PeerOpinionGiver
+}
+
+// OpinionGivers is a map of OpinionGiver.
+type OpinionGivers map[identity.ID]OpinionGiver
+
+// Query retrieves the opinions about the given conflicts and timestamps.
+func (o *OpinionGiver) Query(ctx context.Context, conflictIDs []string, timestampIDs []string) (opinions opinion.Opinions, err error) {
+	for i := 0; i < StatementParameters.WaitForStatement; i++ {
+		if o.view != nil {
+			opinions, err = o.view.Query(ctx, conflictIDs, timestampIDs)
+			if err == nil {
+				return opinions, nil
+			}
+		}
+		time.Sleep(time.Second)
+	}
+
+	return o.pog.Query(ctx, conflictIDs, timestampIDs)
+}
+
+// ID returns the identifier of the underlying Peer.
+func (o *OpinionGiver) ID() identity.ID {
+	return o.id
+}
+
+// OpinionGiverFunc returns a slice of opinion givers.
+func OpinionGiverFunc() (givers []opinion.OpinionGiver, err error) {
+	opinionGiversMap := make(map[identity.ID]*OpinionGiver)
+	opinionGivers := make([]opinion.OpinionGiver, 0)
+
+	for _, v := range Registry().NodesView() {
+		opinionGiversMap[v.ID()] = &OpinionGiver{
+			id:   v.ID(),
+			view: v,
+		}
+	}
+
+	for _, p := range autopeering.Discovery().GetVerifiedPeers() {
+		fpcService := p.Services().Get(service.FPCKey)
+		if fpcService == nil {
+			continue
+		}
+		if _, ok := opinionGiversMap[p.ID()]; !ok {
+			opinionGiversMap[p.ID()] = &OpinionGiver{
+				id:   p.ID(),
+				view: nil,
+			}
+		}
+		opinionGiversMap[p.ID()].pog = &PeerOpinionGiver{p: p}
+	}
+
+	for _, v := range opinionGiversMap {
+		opinionGivers = append(opinionGivers, v)
+	}
+
+	return opinionGivers, nil
+}
+
+// endregion /////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region PeerOpinionGiver /////////////////////////////////////////////////////////////////////////////////////////////
+
+// PeerOpinionGiver implements the OpinionGiver interface based on a peer.
+type PeerOpinionGiver struct {
+	p *peer.Peer
+}
+
+// Query queries another node for its opinion.
+func (pog *PeerOpinionGiver) Query(ctx context.Context, conflictIDs []string, timestampIDs []string) (opinion.Opinions, error) {
+	if pog == nil {
+		return nil, fmt.Errorf("unable to query opinions, PeerOpinionGiver is nil")
+	}
+
+	var opts []grpc.DialOption
+	opts = append(opts, grpc.WithInsecure())
+
+	// connect to the FPC service
+	conn, err := grpc.Dial(pog.Address(), opts...)
+	if err != nil {
+		return nil, fmt.Errorf("unable to connect to FPC service: %w", err)
+	}
+	defer conn.Close()
+
+	client := votenet.NewVoterQueryClient(conn)
+	query := &votenet.QueryRequest{ConflictIDs: conflictIDs, TimestampIDs: timestampIDs}
+	reply, err := client.Opinion(ctx, query)
+	if err != nil {
+		metrics.Events().QueryReplyError.Trigger(&metrics.QueryReplyErrorEvent{
+			ID:           pog.p.ID().String(),
+			OpinionCount: len(conflictIDs) + len(timestampIDs),
+		})
+		return nil, fmt.Errorf("unable to query opinions: %w", err)
+	}
+
+	metrics.Events().FPCInboundBytes.Trigger(uint64(proto.Size(reply)))
+	metrics.Events().FPCOutboundBytes.Trigger(uint64(proto.Size(query)))
+
+	// convert int32s in reply to opinions
+	opinions := make(opinion.Opinions, len(reply.Opinion))
+	for i, intOpn := range reply.Opinion {
+		opinions[i] = opinion.ConvertInt32Opinion(intOpn)
+	}
+
+	return opinions, nil
+}
+
+// ID returns the identifier of the underlying Peer.
+func (pog *PeerOpinionGiver) ID() identity.ID {
+	return pog.p.ID()
+}
+
+// Address returns the FPC address of the underlying Peer.
+func (pog *PeerOpinionGiver) Address() string {
+	fpcServicePort := pog.p.Services().Get(service.FPCKey).Port()
+	return net.JoinHostPort(pog.p.IP().String(), strconv.Itoa(fpcServicePort))
+}
+
+// endregion /////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region OpinionRetriever /////////////////////////////////////////////////////////////////////////////////////////////
+
+// OpinionRetriever returns the current opinion of the given id.
+func OpinionRetriever(id string, objectType vote.ObjectType) opinion.Opinion {
+	switch objectType {
+	case vote.TimestampType:
+		// TODO: implement
+		return opinion.Like
+	default: // conflict type
+		transactionID, err := ledgerstate.TransactionIDFromBase58(id)
+		if err != nil {
+			plugin.LogErrorf("received invalid vote request for branch '%s'", id)
+
+			return opinion.Unknown
+		}
+
+		opinionEssence := ConsensusMechanism().TransactionOpinionEssence(transactionID)
+
+		if opinionEssence.LevelOfKnowledge() == fcob.Pending {
+			return opinion.Unknown
+		}
+
+		if !opinionEssence.Liked() {
+			return opinion.Dislike
+		}
+
+		return opinion.Like
+	}
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region RemoteLogger /////////////////////////////////////////////////////////////////////////////////////////////////
+
+const (
+	remoteLogType = "statement"
+)
+
+var (
+	remoteLogger *remotelog.RemoteLoggerConn
+	myID         string
+	clockEnabled bool
+)
+
+func configureRemoteLogger() {
+	remoteLogger = remotelog.RemoteLogger()
+
+	if local.GetInstance() != nil {
+		myID = local.GetInstance().ID().String()
+	}
+
+	clockEnabled = !node.IsSkipped(clock.Plugin())
+}
+
+func sendToRemoteLog(msgID, issuerID string, issuedTime, arrivalTime, solidTime int64) {
+	m := statementLog{
+		NodeID:       myID,
+		MsgID:        msgID,
+		IssuerID:     issuerID,
+		IssuedTime:   issuedTime,
+		ArrivalTime:  arrivalTime,
+		SolidTime:    solidTime,
+		DeltaArrival: arrivalTime - issuedTime,
+		DeltaSolid:   solidTime - issuedTime,
+		Clock:        clockEnabled,
+		Sync:         Tangle().Synced(),
+		Type:         remoteLogType,
+	}
+	_ = remoteLogger.Send(m)
+}
+
+type statementLog struct {
+	NodeID       string `json:"nodeID"`
+	MsgID        string `json:"msgID"`
+	IssuerID     string `json:"issuerID"`
+	IssuedTime   int64  `json:"issuedTime"`
+	ArrivalTime  int64  `json:"arrivalTime"`
+	SolidTime    int64  `json:"solidTime"`
+	DeltaArrival int64  `json:"deltaArrival"`
+	DeltaSolid   int64  `json:"deltaSolid"`
+	Clock        bool   `json:"clock"`
+	Sync         bool   `json:"sync"`
+	Type         string `json:"type"`
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region Statement ////////////////////////////////////////////////////////////////////////////////////////////////////
+
+func makeStatement(roundStats *vote.RoundStats) {
+	// TODO: add check for Mana threshold
+
+	timestamps := statement.Timestamps{}
+	conflicts := statement.Conflicts{}
+
+	for id, v := range roundStats.ActiveVoteContexts {
+		switch v.Type {
+		case vote.TimestampType:
+			messageID, err := tangle.NewMessageID(id)
+			if err != nil {
+				// TODO
+				break
+			}
+			timestamps = append(timestamps, statement.Timestamp{
+				ID: messageID,
+				Opinion: statement.Opinion{
+					Value: v.LastOpinion(),
+					Round: uint8(v.Rounds)}},
+			)
+		case vote.ConflictType:
+			messageID, err := ledgerstate.TransactionIDFromBase58(id)
+			if err != nil {
+				// TODO
+				break
+			}
+			conflicts = append(conflicts, statement.Conflict{
+				ID: messageID,
+				Opinion: statement.Opinion{
+					Value: v.LastOpinion(),
+					Round: uint8(v.Rounds)}},
+			)
+		default:
+		}
+	}
+
+	broadcastStatement(conflicts, timestamps)
+}
+
+// broadcastStatement broadcasts a statement via communication layer.
+func broadcastStatement(conflicts statement.Conflicts, timestamps statement.Timestamps) {
+	msg, err := Tangle().IssuePayload(statement.New(conflicts, timestamps))
+
+	if err != nil {
+		plugin.LogWarnf("error issuing statement: %s", err)
+		return
+	}
+
+	plugin.LogDebugf("issued statement %s", msg.ID())
+}
+
+func readStatement(messageID tangle.MessageID) {
+	Tangle().Storage.Message(messageID).Consume(func(msg *tangle.Message) {
+		messagePayload := msg.Payload()
+		if messagePayload.Type() != statement.StatementType {
+			return
+		}
+		statementPayload, ok := messagePayload.(*statement.Statement)
+		if !ok {
+			plugin.LogDebug("could not cast payload to statement object")
+			return
+		}
+
+		// TODO: check if the Mana threshold of the issuer is ok
+
+		issuerID := identity.NewID(msg.IssuerPublicKey())
+		// Skip ourselves
+		if issuerID == local.GetInstance().ID() {
+			return
+		}
+
+		issuerRegistry := Registry().NodeView(issuerID)
+
+		issuerRegistry.AddConflicts(statementPayload.Conflicts)
+
+		issuerRegistry.AddTimestamps(statementPayload.Timestamps)
+
+		Tangle().Storage.MessageMetadata(messageID).Consume(func(messageMetadata *tangle.MessageMetadata) {
+			sendToRemoteLog(
+				msg.ID().String(),
+				issuerID.String(),
+				msg.IssuingTime().UnixNano(),
+				messageMetadata.ReceivedTime().UnixNano(),
+				messageMetadata.SolidificationTime().UnixNano(),
+			)
+		})
+	})
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
