+++ plugins/messagelayer/parameters.go
@@ -1,9 +1,78 @@
 package messagelayer
 
-import flag "github.com/spf13/pflag"
+import "github.com/iotaledger/hive.go/configuration"
+
+// Parameters contains the configuration parameters used by the message layer.
+var Parameters = struct {
+	// TangleWidth can be used to specify the number of tips the Tangle tries to maintain.
+	TangleWidth int `default:"0" usage:"the width of the Tangle"`
+
+	// Snapshot contains snapshots related configuration parameters.
+	Snapshot struct {
+		// File is the path to the snapshot file.
+		File string `default:"./snapshot.bin" usage:"the path to the snapshot file"`
+	}
+
+	// FCOB contains parameters related to the fast consensus of barcelona.
+	FCOB struct {
+		AverageNetworkDelay int `default:"5" usage:"the avg. network delay to use for FCoB rules"`
+	}
+}{}
+
+// FPCParameters contains the configuration parameters used by the FPC consensus.
+var FPCParameters = struct {
+	// BindAddress defines on which address the FPC service should listen.
+	BindAddress string `default:"0.0.0.0:10895" usage:"the bind address on which the FPC vote server binds to"`
+
+	// Listen defines if the FPC service should listen.
+	Listen bool `default:"true" usage:"if the FPC service should listen"`
+
+	// RoundInterval defines how long a round lasts (in seconds).
+	RoundInterval int64 `default:"10" usage:"FPC round interval [s]"`
+
+	// QuerySampleSize defines how many nodes will be queried each round.
+	QuerySampleSize int `default:"21" usage:"Size of the voting quorum (k)"`
+}{}
+
+// StatementParameters contains the configuration parameters used by the FPC statements in the tangle.
+var StatementParameters = struct {
+	// WaitForStatement is the time in seconds for which the node wait for receiving the new statement.
+	WaitForStatement int `default:"5" usage:"the time in seconds for which the node wait for receiving the new statement"`
+
+	// WriteStatement defines if the node should write statements.
+	WriteStatement bool `default:"false" usage:"if the node should make statements"`
+
+	// ManaThreshold defines the Mana threshold to accept/write a statement.
+	ManaThreshold float64 `default:"1" usage:"Mana threshold to accept/write a statement"`
+
+	// CleanInterval defines the time interval [in minutes] for cleaning the statement registry.
+	CleanInterval int `default:"5" usage:"the time in minutes after which the node cleans the statement registry"`
+
+	// DeleteAfter defines the time [in minutes] after which older statements are deleted from the registry.
+	DeleteAfter int `default:"5" usage:"the time in minutes after which older statements are deleted from the registry"`
+}{}
+
+// SyncBeaconFollowerParameters contains the configuration parameters used by the syncbeacon follower plugin.
+var SyncBeaconFollowerParameters = struct {
+	// FollowNodes defines the list of nodes this node should follow to determine its sync status.
+	FollowNodes []string `default:"Gm7W191NDnqyF7KJycZqK7V6ENLwqxTwoKQN4SmpkB24,9DB3j9cWYSuEEtkvanrzqkzCQMdH1FGv3TawJdVbDxkd" usage:"list of trusted nodes to follow their sync status"`
+
+	// MaxTimeWindowSec defines the maximum time window for which a sync payload would be considerable.
+	MaxTimeWindowSec int `default:"10" usage:"the maximum time window for which a sync payload would be considerable"`
+
+	// MaxTimeOffline defines the maximum time a beacon node can stay without receiving updates.
+	MaxTimeOffline int `default:"70" usage:"the maximum time the node should stay synced without receiving updates"`
+
+	// CleanupInterval defines the interval that old beacon status are cleaned up.
+	CleanupInterval int `default:"10" usage:"the interval at which cleanups are done"`
+
+	// SyncPercentage defines the percentage of following nodes that have to be synced.
+	SyncPercentage float64 `default:"0.5" usage:"percentage of nodes being followed that need to be synced in order to consider the node synced"`
+}{}
 
 func init() {
-	flag.String(CfgMessageLayerSnapshotFile, "./snapshot.bin", "the path to the snapshot file")
-	flag.Int(CfgMessageLayerFCOBAverageNetworkDelay, 5, "the avg. network delay to use for FCoB rules")
-	flag.Int(CfgTangleWidth, 0, "the width of the Tangle")
+	configuration.BindParameters(&Parameters, "messageLayer")
+	configuration.BindParameters(&FPCParameters, "fpc")
+	configuration.BindParameters(&StatementParameters, "statement")
+	configuration.BindParameters(&SyncBeaconFollowerParameters, "syncbeaconfollower")
 }
